Partially fix circular import
