+++ saleor/webhook/payload_helpers.py
@@ -1,0 +1,38 @@
+from typing import TYPE_CHECKING, Any, Optional
+
+import graphene
+from django.utils import timezone
+from graphene.utils.str_converters import to_camel_case
+
+from .. import __version__
+from ..account.models import User
+
+if TYPE_CHECKING:
+    from ..plugins.base_plugin import RequestorOrLazyObject
+
+
+def generate_requestor(requestor: Optional["RequestorOrLazyObject"] = None):
+    if not requestor:
+        return {"id": None, "type": None}
+    if isinstance(requestor, User):
+        return {"id": graphene.Node.to_global_id("User", requestor.id), "type": "user"}
+    return {"id": requestor.name, "type": "app"}  # type: ignore[union-attr]
+
+
+def generate_meta(*, requestor_data: dict[str, Any], camel_case=False, **kwargs):
+    meta_result = {
+        "issued_at": timezone.now().isoformat(),
+        "version": __version__,
+        "issuing_principal": requestor_data,
+    }
+
+    meta_result.update(kwargs)
+
+    if camel_case:
+        meta = {}
+        for key, value in meta_result.items():
+            meta[to_camel_case(key)] = value
+    else:
+        meta = meta_result
+
+    return meta
